Create compareGeometricGrowthOnDiceAverages
uses a geometric growth rate, showing the dice-rolls appropriate for the resultant averages.

Used for investigating an appropriate rate to adjust numbers by with a risk-reward system for balancing Damage Vs. Accuracy.

First number is the initial percentage change (set to .14 here) and  the number of terms can be expanded to lengthen the results. Given that it's geometric, distant results become increasing outliers that would break a TTRPG game balance.

1.5 looks appropriate for a familiar d20 system